requirements.txt
pygame-ce
json5